datetime: strptime: fix '%s %z'
This format is used by Git to store commit dates, and it is a sensible
way to represent them in a simple manner.

It works correctly in C and perl, but in Ruby it returns the wrong date:

  require 'date'
  date = DateTime.strptime('1377283786 -0700', '%s %z')
  puts date.strftime('%s %z')

The fix is simple and don't affect the normal usage of '%s'.

For more information, see bug:

https://bugs.ruby-lang.org/issues/7445

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>